\name{PepsNMRData-package}
\alias{PepsNMRData-package}
\alias{PepsNMRData}
\docType{package}
\title{Raw and (partially) pre-processed metabolomic datasets supporting the \code{PepsNMR} package
}
\description{
This package provides RData and Bruker files of raw and (partially) pre-processed metabolomic 1HNMR datasets from human urine and serum.


\describe{
  \item{\code{\link{Data_HS_sp}}}{}
  \item{\code{\link{Data_HU_sp}}}{}
  \item{\code{\link{FidData_HS}}}{}
  \item{\code{\link{FidData_HU}}}{}
  \item{\code{\link{FidInfo_HS_sp}}}{}
  \item{\code{\link{FidInfo_HS}}}{}
  \item{\code{\link{FidInfo_HU_sp}}}{}
  \item{\code{\link{FidInfo_HU}}}{}
  \item{\code{\link{FinalSpecta_HS}}}{}
  \item{\code{\link{FinalSpecta_HU}}}{}
  \item{\code{\link{Group_HS}}}{}
  \item{\code{\link{Group_HU}}}{}
}
}
\details{
\tabular{ll}{
Package: \tab PepsNMRData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
License: \tab GPLv2\cr
}

2 different datasets: HUrine and HSerum

}
\author{
Benoît Legat, Bernadette Govaerts & Manon Martin

Maintainer: Manon Martin <manon.martin@uclouvain.be>
}
\references{
PepsNMR article

  Rousseau, R. (2011). \emph{Statistical contribution to the analysis of metabonomics data in 1H NMR spectroscopy}
  (Doctoral dissertation, PhD thesis. Institut de statistique, biostatistique et sciences actuarielles, Université catholique de Louvain, Belgium).
}
\keyword{ package }

\examples{

}
